requirements.txt
# Unified XAI Platform - Dependencies

# Core ML/AI
tensorflow>=2.10.0
numpy>=1.21.0

# Audio Processing
librosa>=0.10.0

# Image Processing
opencv-python>=4.5.0
Pillow>=9.0.0
scikit-image>=0.19.0

# PyTorch & TorchXRayVision (Image Model)
torch>=2.0.0
torchvision>=0.15.0
torchxrayvision>=1.2.0
pytorch-grad-cam>=1.4.0

# XAI Libraries
lime>=0.2.0
shap>=0.41.0

# Web Interface
streamlit>=1.25.0

# Utilities
scipy>=1.7.0
